.github/workflows/helpers/requirements.txt
# requirements.txt

# HTTP requests for GitHub API
requests>=2.25.1

# Google Generative AI SDK
google-generativeai>=0.3.2
